requirements.txt
python-telegram-bot==13.12
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.4
yandex_tracker_client==1.1
fastapi[all]==0.115.6
